requirements.txt
python-telegram-bot
python-dotenv
gspread
google-auth
APScheduler
pytz
